feat: add setting nav link
