+++ src/components/settings/FileList.tsx
@@ -1,0 +1,6 @@
+import React from 'react'
+
+
+export const FileListSettingView = () => (
+  <div>FileListSettingView</div>
+)
