+++ src/components/settings/index.tsx
@@ -1,61 +1,37 @@
 import styled from '@emotion/styled'
-import React, { useState } from 'react'
-import { useDispatch } from 'react-redux'
-import { Link } from 'react-router-dom'
+import React from 'react'
+import { NavLink,Route, Switch } from 'react-router-dom'
 
-import { settingActions } from '@/features/settingSlice'
+import { FileListSettingView } from './FileList'
+import { TokenSettingsView } from './Token'
 
-const Layout = styled.div`
-  margin: 10px;
-  width: 280px;
+const StyledNavLink = styled(NavLink)`
+  &.active {
+    color: #18CD8C;
+  }
 `
 
-const Header = styled.header`
+const StickyNav = styled.header`
   display: flex;
+  justify-content: space-around;
 `
-const Head = styled.h1`
-  font-weight: 700;
-  margin: 10px 0;
-`
-const CloseButton = styled(Link)`
-  color: gray;
-`
-const TextInputField = styled.input`
-  display: block;
-  width: 100%;
-`
-const SaveButton = styled.button`
-  padding: 8px 0px;
-  margin-top: 10px;
-  width: 100%;
-  background: rgba(24,205,140,0.1);
-  border-radius: 8px;
-  color: #18CD8C;
-  font-weight: bold;
-  font-size: 14px;
-`
+
 // https://docs.github.com/en/rest/reference/pulls#list-pull-requests-files
 export const SettingsView = () => {
-  const dispatch = useDispatch()
-  const [token, setToken] = useState('')
-
-  const saveToken = () => {
-    dispatch(settingActions.saveToken(token))
-  }
-
   return (
-    <Layout>
-      <Header>
-        <Head>Add GitHub API Token</Head>
-        <CloseButton to='/'>X</CloseButton>
-      </Header>
-      <TextInputField
-        value={token}
-        onChange={({ target }: React.ChangeEvent<HTMLInputElement>) =>
-          setToken(target.value)
-        }
-      />
-      <SaveButton onClick={saveToken}>Save</SaveButton>
-    </Layout>
+    <>
+      <StickyNav>
+        <StyledNavLink activeClassName='active' to='/settings/file-list'>
+          Files
+        </StyledNavLink >
+        <StyledNavLink activeClassName='active' to='/settings/token'>
+          Token
+        </StyledNavLink >
+      </StickyNav>
+      <Switch>
+        <Route path='/settings/token' component={TokenSettingsView} />
+        <Route path='/settings/file-list' component={FileListSettingView} />
+      </Switch>
+    </>
   )
 }
